chunkをallに変更

+++ Sources/App/Models/logMemoryUsage.swift
@@ -6,9 +6,9 @@ func logMemoryUsage(title: String, logger: Logger) {
     do {
       let contents = try String(contentsOfFile: "/proc/self/statm", encoding: .utf8)
       let parts = contents.split(separator: " ")
-      if let pages = parts.first.map(String.init).flatMap(Int.init) {
-        let pageSize: Int = 4096
-        logger.critical("\(title),MEMORY: \(pages * pageSize)")
+      if let pages = parts.first.map(String.init).flatMap(UInt64.init) {
+        let pageSize: UInt64 = 4096
+        logger.critical("\(title),MEMORY: \(pages * pageSize)") // Bytes
       }
     } catch {
       logger.error("\(title),MEMORY-ERR: \(error)")
